Feat: Implemented listing feature according to the grouping in the sidebar
