+++ client/package.json
@@ -13,7 +13,8 @@
     "axios": "^1.7.8",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.4.0"
+    "react-icons": "^5.4.0",
+    "react-spinners": "^0.14.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
