+++ client/src/Components/Hero.jsx
@@ -1,0 +1,110 @@
+import React from "react";
+
+const Hero = () => {
+  return (
+    <div className="hero">
+      <section>
+        <h2>New Releases</h2>
+        <div className="tracks">
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/1.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/2.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/3.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/4.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/5.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+        </div>
+      </section>
+      <section>
+        <h2>Top Hits</h2>
+        <div className="tracks">
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/6.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/7.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/8.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/9.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+          <div className="track-card">
+            <div className="image-container">
+              <img src="./src/assets/photos/10.jpg" alt="" />
+            </div>
+            <div className="track-info">
+              <h3>Track Name</h3>
+              <p>Artist Name</p>
+            </div>
+          </div>
+        </div>
+      </section>
+    </div>
+  );
+};
+
+export default Hero;
