+++ client/src/styles/list.css
@@ -1,0 +1,82 @@
+.list {
+  height: 90vh;
+  overflow: hidden;
+  background-color: black;
+  transition: width 0.3s ease;
+  overflow-y: scroll;
+  scrollbar-width: none;
+}
+
+.list h2 {
+  font-size: 30px;
+  font-weight: 700;
+  text-align: center;
+  padding: 10px;
+  color: white;
+}
+
+ul.items-list {
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  flex-direction: column;
+  width: 100vw;
+}
+
+ul.items-list li {
+  width: 90%;
+  padding: 10px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  cursor: pointer;
+  margin: 10px;
+  border-radius: 5px;
+  color: white;
+  transition: background 0.3s ease;
+}
+
+ul.items-list li:hover {
+  background-color: rgba(128, 128, 128, 0.1);
+}
+
+ul.items-list p.rank {
+  font-size: 20px;
+  font-weight: 700;
+}
+
+ul.items-list li .image-container {
+  width: 100px;
+  height: 100px;
+  border-radius: 5px;
+  overflow: hidden;
+  margin: 0 20px;
+}
+
+ul.items-list li .image-container img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  cursor: pointer;
+  transition: transform 0.1s ease;
+}
+
+ul.items-list li .image-container img:hover {
+  transform: scale(1.1);
+}
+
+.item-info {
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+  flex-direction: column;
+}
+
+.item-info h3 {
+  font-size: 18px;
+  font-weight: 700;
+}
+
+.item-info p {
+  font-size: 14px;
+}
